migrate all openshift legacy commands to catalog
