+++ ai-services/internal/pkg/application/podman/info.go
@@ -1,18 +1,34 @@
 package podman
 
 import (
+	"bytes"
 	"fmt"
+	"strings"
+	"text/template"
 
 	"github.com/project-ai-services/ai-services/assets"
 	"github.com/project-ai-services/ai-services/internal/pkg/application/types"
+	"github.com/project-ai-services/ai-services/internal/pkg/catalog"
+	catalogClient "github.com/project-ai-services/ai-services/internal/pkg/catalog/client"
+	catalogTypes "github.com/project-ai-services/ai-services/internal/pkg/catalog/types"
 	"github.com/project-ai-services/ai-services/internal/pkg/cli/helpers"
 	"github.com/project-ai-services/ai-services/internal/pkg/cli/templates"
+	cliUtils "github.com/project-ai-services/ai-services/internal/pkg/cli/utils"
 	"github.com/project-ai-services/ai-services/internal/pkg/logger"
+	"github.com/project-ai-services/ai-services/internal/pkg/runtime"
 	"github.com/project-ai-services/ai-services/internal/pkg/vars"
 )
 
 // Info displays detailed information about an application.
 func (p *PodmanApplication) Info(opts types.InfoOptions) error {
+	if opts.Legacy {
+		return legacyInfo(opts, p.runtime)
+	} else {
+		return renderApplicationInfo(opts.Name)
+	}
+}
+
+func legacyInfo(opts types.InfoOptions, runtime runtime.Runtime) error {
 	// Step1: Do List pods and filter for given application name
 
 	listFilters := map[string][]string{}
@@ -20,7 +36,7 @@ func (p *PodmanApplication) Info(opts types.InfoOptions) error {
 		listFilters["label"] = []string{fmt.Sprintf("ai-services.io/application=%s", opts.Name)}
 	}
 
-	pods, err := p.runtime.ListPods(listFilters)
+	pods, err := runtime.ListPods(listFilters)
 	if err != nil {
 		return fmt.Errorf("failed to list pods: %w", err)
 	}
@@ -45,12 +61,133 @@ func (p *PodmanApplication) Info(opts types.InfoOptions) error {
 	// Step3: Read and print the info.md file
 	tp := templates.NewEmbedTemplateProvider(&assets.ApplicationFS)
 
-	if err := helpers.PrintInfo(tp, p.runtime, opts.Name, appTemplate); err != nil {
+	if err := helpers.PrintInfo(tp, runtime, opts.Name, appTemplate); err != nil {
 		// not failing if overall info command, if we cannot display Info
 		logger.Errorf("failed to display info: %v\n", err)
 
 		return nil
 	}
+
+	return nil
+}
+
+func renderApplicationInfo(appName string) error {
+	appClient, err := catalogClient.NewApplicationClient()
+	if err != nil {
+		return fmt.Errorf("failed to create application client: %w", err)
+	}
+
+	app, err := cliUtils.GetAppByName(appClient, appName)
+	if err != nil {
+		if strings.Contains(err.Error(), "not found") {
+			logger.Warningf("Application: '%s' does not exist", appName)
+
+			return nil
+		}
+
+		return err
+	}
+
+	application, err := appClient.GetApplication(app.ID)
+	if err != nil {
+		return fmt.Errorf("failed to get application: %w", err)
+	}
+
+	appPS, err := appClient.GetApplicationPS(app.ID)
+	if err != nil {
+		return fmt.Errorf("failed to get application pods: %w", err)
+	}
+
+	logger.Infoln("Application Name: " + application.Name)
+	logger.Infoln("Application Template: " + application.CatalogID)
+	logger.Infoln("Application Version: " + application.Version)
+
+	return printServicesInfo(application.Services, appPS)
+}
+
+func printServicesInfo(services []catalogTypes.ApplicationService, appPS *catalogTypes.ApplicationPSResponse) error {
+	catalogProvider, err := catalog.NewCatalogProvider()
+	if err != nil {
+		return fmt.Errorf("failed to create catalog provider: %w", err)
+	}
+
+	logger.Infoln("Info:")
+	logger.Infoln("-------")
+	logger.Infoln("Day N: ")
+
+	for _, service := range services {
+		params := map[string]string{}
+		params["SERVICE_NAME"] = service.Type
+
+		uiStatus, apiStatus := getContainerStatus(appPS.Services, service.CatalogID)
+		params["UI_STATUS"] = uiStatus
+		params["API_STATUS"] = apiStatus
+
+		for _, endpoint := range service.Endpoints {
+			urlType, urlTypeOk := endpoint["type"].(string)
+			url, urlOk := endpoint["url"].(string)
+			if urlTypeOk && urlOk {
+				params[strings.ToUpper(urlType)+"_URL"] = url
+			}
+		}
+
+		tmpls, err := catalogProvider.LoadServicesMD(service.CatalogID)
+		if err != nil {
+			return fmt.Errorf("failed to load service md files: %w", err)
+		}
+
+		err = printInfo(tmpls, params)
+		if err != nil {
+			return fmt.Errorf("failed to load application info: %w", err)
+		}
+	}
+
+	return nil
+}
+
+func getContainerStatus(services []catalogTypes.Pod, catalogID string) (string, string) {
+	uiStatus, apiStatus := "", ""
+
+	for _, servicePod := range services {
+		if strings.HasPrefix(servicePod.PodName, catalogID) {
+			for _, podContainer := range servicePod.Containers {
+				// TODO: Set the container status in info.md generically
+				uiContainerName := fmt.Sprintf("%s-ui", servicePod.PodName)
+				apiContainerName := ""
+				if strings.Contains(podContainer.Name, "backend-server") {
+					apiContainerName = podContainer.Name
+				} else {
+					apiContainerName = fmt.Sprintf("%s-%s-api", servicePod.PodName, catalogID)
+				}
+
+				if podContainer.Name == uiContainerName && podContainer.Healthy {
+					uiStatus = "running"
+				}
+				if podContainer.Name == apiContainerName && podContainer.Healthy {
+					apiStatus = "running"
+				}
+			}
+		}
+	}
+
+	return uiStatus, apiStatus
+}
+
+func printInfo(tmpls map[string]*template.Template, params map[string]string) error {
+	tmpl, ok := tmpls["info.md"]
+	if !ok {
+		logger.Warningf("failed to find info.md template")
+
+		return nil
+	}
+
+	var rendered bytes.Buffer
+	if err := tmpl.Execute(&rendered, params); err != nil {
+		return fmt.Errorf("failed to execute info.md: %w", err)
+	}
+	value := rendered.String()
+	value = strings.ReplaceAll(value, "Day N:\n", "")
+	logger.Infoln(value)
 
 	return nil
 }
